- Removed unused imports from previous QuickLook implementation. - Removed one unneded feature key (canSave). - Renamed DisplayViewProtocol to DisplayExtension. - Make the default initialization of previews as a protocol extension.
