+++ ownCloud/Viewer/WebViewDisplayViewController.swift
@@ -10,30 +10,15 @@ import UIKit
 import ownCloudSDK
 import WebKit
 
-class WebViewDisplayViewController: DisplayViewController, DisplayViewProtocol {
+class WebViewDisplayViewController: DisplayViewController, DisplayExtension {
+	
+	static var displayExtensionIdentifier: String = "imageViewer"
 	static var supportedMimeTypes: [String] =
 		["image/jpeg",
 		 "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
 		 "application/vnd.openxmlformats-officedocument.wordprocessingml.document"]
 
 	static var features: [String : Any]? = [FeatureKeys.canEdit : true]
-
-	static func webViewExtension(identifier: String) -> OCExtension {
-		let rawIdentifier: OCExtensionIdentifier =  OCExtensionIdentifier(rawValue: identifier)
-		var locationIdentifiers: [OCExtensionLocationIdentifier] = []
-		for mimeType in supportedMimeTypes {
-			let locationIdentifier: OCExtensionLocationIdentifier = OCExtensionLocationIdentifier(rawValue: mimeType)
-			locationIdentifiers.append(locationIdentifier)
-		}
-
-		let features: [String : Any] = WebViewDisplayViewController.features!
-
-		let webViewExtension = OCExtension(identifier: rawIdentifier, type: .viewer, locations: locationIdentifiers, features: features) { (_ rawExtension, _ context, _ error) -> Any? in
-			return WebViewDisplayViewController()
-		}
-
-		return webViewExtension!
-	}
 
 	var webView: WKWebView?
